Rename existing bloom filter, start bloom filter sub module
