+++ src/gadgets.rs
@@ -1,6 +1,5 @@
-mod bit_selector;
+mod bloom_filter_old;
 mod bloom_filter;
-mod byte_selector;
 mod hash;
 mod response_accumulator;
 pub mod wnn;
